% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird_counts_data.R
\docType{data}
\name{bird_counts}
\alias{bird_counts}
\title{The bird_counts dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14925 rows and 6 columns.
}
\source{
https://raw.githubusercontent.com/rfordatascience/tidytuesday/master/data/2019/2019-06-18/bird_counts.csv
}
\usage{
bird_counts
}
\description{
This data shows the numerous bird species counted by the birdwatchers in Hamilton area of Ontario, every year since Christmas 1921.
}
\keyword{datasets}
