% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_data.R
\docType{data}
\name{hourly_data}
\alias{hourly_data}
\title{The hourly_data data set}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 61 rows and 3 columns.
}
\usage{
hourly_data
}
\description{
This data set is a subset of the bird_counts data set and it shows an approximate range of hours required to count birds. This data was created for the function count_by_hours.
}
\keyword{datasets}
