% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbirds.pkg-package.R
\docType{package}
\name{rbirds.pkg-package}
\alias{rbirds.pkg}
\alias{rbirds.pkg-package}
\title{rbirds.pkg: Birds counted in Ontario during Christmas.}
\description{
This packages contains data about birds counted in Christmas in Ontario from 1921 to 2017. The data is retrieved from the tidytuesday project on GitHub. This package allows users to explore the data with various functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://etc5523-2022.github.io/rpkg-Adarsh098/}
}

}
\author{
\strong{Maintainer}: Adarsh More \email{amor0060@student.monash.edu}

}
\keyword{internal}
