% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_five_birds_yearly.R
\name{top_five_birds_yearly}
\alias{top_five_birds_yearly}
\title{Top five birds counted yearly}
\usage{
top_five_birds_yearly(num)
}
\arguments{
\item{num}{A numeric input of year between 1921 - 2017.}
}
\description{
This function takes year as an input and allows user to view the top five birds counted in that year. It is also a parameterised version of the output of panel four of the shiny application of this package where the input "year" corresponds to the argument "num" in the function.
}
\examples{
top_five_birds_yearly(1960)
}
