% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_data.R
\docType{data}
\name{year_data}
\alias{year_data}
\title{The year_data data set}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 75 rows and 3 columns.
}
\usage{
year_data
}
\description{
This data set is a subset of the bird_counts data set and it shows the most counted bird in each year. This data was created for the function yearly_count.
}
\keyword{datasets}
