% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_count.R
\name{yearly_count}
\alias{yearly_count}
\title{Birds counted between 1960-1965 during Christmas.}
\usage{
yearly_count(year)
}
\description{
A function that takes year between 1960-1965 as an input and allows the user to view the three most counted birds in that year. The data set used for this function is the yearly_data.
}
