% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_data.R
\docType{data}
\name{yearly_data}
\alias{yearly_data}
\title{The yearly_data data set}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 375 rows and 3 columns.
}
\usage{
yearly_data
}
\description{
This data set is a subset of the bird_counts data set and it shows the top five most counted birds in each year. This data was created for the function top_five_birds_yearly. This data is also used for the parameterised version of the plot in panel 4 of the shiny application.
}
\keyword{datasets}
