% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_by_hours.R
\name{count_by_hours}
\alias{count_by_hours}
\title{Birds counted hourly.}
\usage{
count_by_hours(how_many_counted)
}
\arguments{
\item{how_many_counted}{A numeric input greater than 0}
}
\description{
A function that takes the count of birds as an input and allows the user to view the number of hours it took to count that many birds.
}
\examples{
count_by_hours(5000)
}
